Add docker-compose recipe to deploy the whole product
