+++ deploy/docker-compose/hoopdata/nginx.conf
@@ -1,0 +1,33 @@
+events {
+    worker_connections  1024;
+}
+
+http {
+
+    server {
+        server_name auth.127.0.0.1.nip.io;
+        listen 80;
+        http2 on;
+        location / {
+            grpc_pass grpc://idp:8080;
+            grpc_set_header Host $host;
+        }
+    }
+
+    server {
+        server_name 127.0.0.1.nip.io;
+        listen 80;
+        http2 on;
+        location / {
+                proxy_pass http://gateway:8009;
+        }
+
+        location ~ ^/protobuf.+ {
+                client_max_body_size 0;
+                grpc_read_timeout 30s;
+                grpc_send_timeout 30s;
+                grpc_socket_keepalive on;
+                grpc_pass grpc://gateway:8010;
+        }
+    }
+}
