+++ deploy/docker-compose/hoopdata/terraform/variables.tf
@@ -1,0 +1,3 @@
+variable public_hostname {
+    type = string
+}
