+++ deploy/docker-compose/hoopdata/tls/nginx.conf
@@ -1,0 +1,38 @@
+events {
+    worker_connections  1024;
+}
+
+http {
+    server {
+        server_name auth.192.168.15.48.nip.io;
+        listen 443 ssl;
+        http2 on;
+        ssl_certificate     /etc/certs/server.crt;
+        ssl_certificate_key /etc/certs/server.key;
+        location / {
+            grpc_pass grpc://idp:8080;
+            grpc_set_header Host $host;
+        }
+    }
+
+    server {
+        server_name 192.168.15.48.nip.io;
+        listen 443 ssl;
+        http2 on;
+        ssl_certificate     /etc/certs/server.crt;
+        ssl_certificate_key /etc/certs/server.key;
+
+        location ~ ^/protobuf.+ {
+                client_max_body_size 0;
+                grpc_read_timeout 30s;
+                grpc_send_timeout 30s;
+                grpc_socket_keepalive on;
+                grpc_pass grpc://192.168.15.48.nip.io:8010;
+                break;
+        }
+
+        location / {
+            proxy_pass http://192.168.15.48.nip.io:8009;
+        }
+    }
+}
